coded methods to generate notes from patterns

+++ zyn.common/globals.cpp
@@ -28,7 +28,7 @@ SimpleNote::SimpleNote() : note(0), velocity(0), length(0), channel(0) {}
 SimpleNote::SimpleNote(
     unsigned int n,
     unsigned int v,
-    unsigned int l,
+    float l,
     unsigned int c)
     : note(n),
       velocity(v),
